add Vote filter

+++ database/migrations/2023_07_07_092811_add_avarage_column_to_profiles_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('profiles', function (Blueprint $table) {
-            $table->decimal('average_vote', 5, 2)->nullable();
+            $table->integer('average_vote')->nullable();
         });
     }
